Adds the calculation of Helmholtz free-energy by summing over modes.
Add two functions.
The first one, called get_Nkpoints(directories), gets the total number of
kpoints in each volume.
The second called calc_free_energy(MF_FIT, N_kpoints, PE,temperature_list)
calculates the Helmholtz free-energy by summing over the smoothed modes
contained in the matrix MF_FIT. The function returns a Nt by Nv matrix,
where Nt is the number of temperatures contained in temperature_list (in
Kelvin) and Nv is the number of volumes (number of columns in MF_FIT).
The script assumes that the modes are given in cm^{-1}. The helmholtz
free-energy is given in eV.